refactor: hexToRgb를 삭제하고 기존의 hexToRgba 유틸함수가 객체를 반환하도록 리펙토링
