+++ apps/web/src/lib/hex-to-rgba.ts
@@ -2,7 +2,7 @@ function hexToRgba(hex: string, alpha: number) {
   const r = parseInt(hex.slice(1, 3), 16);
   const g = parseInt(hex.slice(3, 5), 16);
   const b = parseInt(hex.slice(5, 7), 16);
-  return `rgba(${r},${g},${b},${alpha})`;
+  return { r, g, b, a: alpha };
 }
 
 export default hexToRgba;
